+++ CHANGELOG.md
@@ -1,1 +1,8 @@
-Initial release
+Added the flags `-pairs` and `-safepairs`.  
+
+This uses the (safe) pairs from the Google Doc [WH Pairs list - STP Todd](https://docs.google.com/spreadsheets/d/1XWadBbVkbdi5Ub7bFhCcAhqpHiQXBETbeTg644pkTdI/edit#gid=1034827699) 
+
+Usage:
+- First stop the WH bot and close it. 
+- Run AutoCoins with the startup flag: `go-autocoins.exe -pairs`  
+- Start WH bot again.